Makefile: update Git to v2.48.1.gl1
Update to Git v2.48.1.gl1, which contains a couple of backports. All of
the fixes have been merged upstream.

The first set of backports fixes a couple of regressions introduced by
the new reflog migration code. These fixes fix the migration logic
itself as well as an uninitialized-memory read introduced is the
reftable backend. These backports include:

 - ps/reflog-migration-with-logall-fix at a0bea097 (refs: fix migration
   of reflogs respecting "core.logAllRefUpdates", 2025-01-22).
 - kn/reflog-migration-fix at bc67b4ab (reftable: write correct
   max_update_index to header, 2025-01-15).
 - kn/reflog-migration-fix-fix at f11f0a5a (refs/reftable: fix
   uninitialized memory access of max_index, 2025-01-27).

The second set of backports includes fixes for the Meson build system as
well as a backport of the new zlib-ng backend. These backports ensure
that we can start dogfooding the Meson build system and roll out zlib-ng
to production. These backports include:

  - ps/meson-weak-sha1-build at 6a0ee54f (meson: provide a summary of
    configured backends, 2024-12-30).
  - tc/meson-use-our-version-def-h at 4771501c (meson: ensure correct
    version-def.h is used, 2025-01-14).
  - sj/meson-doc-technical-fix at 1dca492e (meson: fix missing deps for
    technical articles, 2025-01-14).
  - ps/build-meson-fixes at 7304bd2b (ci: wire up Visual Studio build
    with Meson, 2025-01-22).
  - ps/zlib-ng at 78cdeed4 (ci: make "linux-musl" job use zlib-ng,
    2025-01-28).