migration: Pass the migrations as parameters
The migrations to be run are set as global variables within the
'migration' package. This flow requires that individual migrations need
to be declared within the package and that all dependencies for
migrations should flow from the manager to the migration.

To make it easier to manager dependencies, let's inverse the flow and
pass the migrations as an argument to the manager. This allows
migrations to be defined with their dependencies already met.

Previously `RecordKeyCreation` and `RecordKeyDeletion` accessed the
global `migrations` variable. Now we introduce a new interface
`migration.StateManager` which provides these functions with internal
access to all migrations.