raft: Fix data race in raft transport layer
The msgs contain entries that share the same underlying raftpb.Message,
the same message is being sent to multiple followers and then the
goroutines will all be working with the same msg.Entries slice. This
leads to a data race because some goroutines will be reading
msg.Entries[i].Data while others are writing to it.