Makefile: Fix building Git with Meson on macOS
When migrating the GDK to build Git with Meson we have hit the following
error on macOS:

    bash: /Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version: No such file or directory
    make: *** [/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version] Error 1

The Makefile is trying to create "_build/deps/git-v2.48/build.version",
but the parent directory does not exist and thus the build rule fails.
This is somewhat puzzling though: what it really should be trying to
build is "_build/deps/git-v2.48.version".

This boils down to a bug in GNU Make. We have the following line in our
Makefile to build Git:

    $(patsubst %,${DEPENDENCY_DIR}/git-\%/build/%,${GIT_EXECUTABLES}): ${DEPENDENCY_DIR}/git-%/Makefile

The first pattern is escaped, and consequently it should not get
replaced by the call to `patsubst`. And it seemingly doesn't, but
somehow Make now decides to treat everything after the escaped pattern
as part of the stem that it shall use. So instead of expanding to the
version (which is what the first, escaped pattern shall match), it
expands to "v2.48/build". Consequently, the following dependencies also
expand to "git-v2.48/build/Makefile" and "git-v2.48/build.version":

    ${DEPENDENCY_DIR}/git-%/Makefile: ${DEPENDENCY_DIR}/git-%.version
    ${DEPENDENCY_DIR}/git-%.version: dependency-version | ${DEPENDENCY_DIR}

The issue only hits on macOS because its default version of GNU Make is
v3.81, which has been released in 2006.

We can work around the issue by avoiding having two patterns in the
first place, one for the version and one for the executable. This is
done by instead only depending on a ".built" file that gets updated
after all executables have been built. Write the list of executables
into the version file to ensure that we build new executables in case
they get appended to `GIT_EXECUTABLES`.