Goroutine leaks: buffering a channel to avoid blocking
When the rate limiter returns an error but the request
context was already cancelled, the `errs` channel has
no listener, which blocks the goroutine that want to
write to it.

By buffering the channel with 1 slot, we can be sure the
error will be written to the channel, which will then be
garbage collected.

Relevant resources:
* https://gitlab.com/gitlab-org/gitaly/-/issues/6618