Makefile: Fix order of bundled Git rules with Make v3.81 and older
In the GNU Make commit f5891a26 (Implement the shortest stem first
search order for pattern-specific variables and pattern rules.,
2009-09-28), the way how pattern rules are being matched has changed in
a backwards-compatible way:

    The pattern-specific variables and pattern rules are now applied in
    the shortest stem first order instead of the definition order
    (variables and rules with the same stem length are still applied in
    the definition order). This produces the usually-desired behavior
    where more specific patterns are preferred. To detect this feature
    search for 'shortest-stem' in the .FEATURES special variable.

Before that commit, pattern rules were evaluated so that the first
matching rule picked. After the change, this was adapted so that the
rule with the shortest stem was matched.

We rely on this behaviour with how we build bundled Git: we've got two
rules for building Git, one that matches "git-%/build/git" and one that
matches "git-%/git", depending on whether we want to use Meson or Make
to build Git. The former rule leads to a shorter stem (the part that the
`%` matches), but is ordered after the other rule. So if we use GNU Make
without the commit, we'd use the wrong rule and thus get the following
broken dependency chain:

     File `build-bundled-git-v2.48' does not exist.
           File `dependency-version' does not exist.
          Must remake target `dependency-version'.
          Successfully remade target file `dependency-version'.
        File `/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build/git-remote-http' does not exist.
         File `/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build/Makefile' does not exist.
          File `/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version' does not exist.
         Must remake target `/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version'.
    bash: /Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version: No such file or directory
    make: *** [/Users/gitlab/builds/gitlab-org/gitlab-development-kit/gitaly/_build/deps/git-v2.48/build.version] Error 1

The commit in question was part of GNU Make v3.82, which is ancient. But
unfortunately, macOS still ships that version by default, so we do hit
the issue in the wild.

Reorder the rules to make them work in modern and ancient versions of
GNU Make.