migration: Close the dry-run manager when the partition is closed
When the partition's `Close()` method is called, we should also close
the dry-run migration manager. Since `combinedMigrationPartition` has an
embedded `storage.Partition` field, any `Close()` operation would
directly be called on that field. We should add an override to ensure
that the dry-run's `Close()` is also called. This would ensure that the
context for the dry-run manager is cancelled and any ongoing operations
are stopped as necessary.