migration: Pass metadata from the request's context
The migration manager uses two context:
1. request context: This is the context provided by the incoming gRPC
request which initiated the migration manager. We use this request to
check if a migration is enabled/disabled.
2. manager's internal context: This context is used to propagate the
partition manager status within the manager. This is the context which
is passed to individual migrations.

Since the manager's internal context doesn't have access to the
featureflags, this is extended to the migrations. A repercussion of
which is that the latest Git version which is behind a feature flag is
not utilized by the migrations since we don't have the required
information in the context.

Let's copy over metadata from the request's context to the manager's
context to solve this issues.