config: Ensure pack-objects cache doesn't point to storage root
We have had a case recently [1] where the pack_objects_cache.dir value
was set to the same path as the Gitaly storage path. The pack-objects
cache is a filesystem cache and is evicted after a customisable age.

This misconfiguration caused repositories to be deleted from the
filesystem when the eviction was triggered.

Validate that the absolute path of pack_objects_cache.dir doesn't match
the absolute path of any configured storage directories, or any of its
parent directories.

[1] https://gitlab.com/gitlab-org/gitaly/-/issues/6611