Implement RaftSnapshotter
RaftSnapshotter will take a snapshot of the partition by packaging
a partition's contents into a tar package that is compressed with lz4.
It creates a tmp file for the snapshot and when its finished the process
with no errors, it will move the file to the snapshot dir configured via
the raft config.