InfoRefs: Only advertize bundle-uri when a bundle exists
In a previous MR, we updated the UploadPack RPC to only
advertize the `bundle-uri` capability when a bundle exists.

https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7572

However, those changes overlooked the `InfoRefs` RPC. This
RPC is also called during a `git clone` operation. We end
up in a situation where, when a bundle DOES NOT exist, the
`InfoRefs` RPC was advertizing the `bundle-uri` capability,
but when the client would call `UploadPack` afterwards to
fetch a bundle, it would fail because the `UploadPack` would
not support the `bundle-uri` capability.

This commit updates the `InfoRefs` RPC to only advertise
the `bundle-uri` capability when a bundle exist. As such,
the behavior of the `InfoRefs` RPC is now on par with the
`UploadPack` RPC, and conflicts such as the one described
above occurs.