repository: Add RPC to dry-run the reftable migrations
To dry-run the reftable migration, we cannot use the migration
frameworks's dry-run functionality since the goroutines is bound to the
partition and the partition is closed when RPCs which triggered the
partition complete.

This means, long-running migrations are cancelled before completion. To
get around this, we want to introduce an RPC who's purpose is to run the
dry-run the reftable migration. The RPC will capture the metrics needed
and finally return an error to ensure that the transaction manager
rollsback the transaction.