Fix `pages` CI job
The current Gitaly CICD pipeline has a `pages`
job that keeps failing. It is an optionnal job
so it is not critically affecting our iterations
but it is still annoying. The job fails because
of a missing dependency when generating protobuf
documentation.

This commit adds the missing `include` statements
into the command to allow the build of the documentation.