Bundle-Uri: Add more logging
The current level of logging for bundle-uri makes it
hard to know if a bundle was generated or not.

This commit adds a bit more logging only when a bundle
is or should be generated. There are no logs being written
when a bundle should not be generated. This is to ensure
we do not log too much, and only when it is relevant.