Bundle-Uri: Activate bundle generation
While everything is in place to support generating
bundles, it is missing the actual call to the
function that will generate those bundles. This
omission was thoughtfully planned because by
enabling it into an isolated commit, it makes it
easiser to revert in case we face some issues.

This commit introduces the call to the function that
will generate bundles, effectively enabling bundle
generation (when the appropriate feature flags are on).

References:
https://gitlab.com/gitlab-org/gitaly/-/issues/5656