raft: Backfill missing fysnc for internal entries
When Raft manager handles internal log entries issued by etcd/raft, it
creates corresponding WAL entries without going through the
TransactionManager. The most crucial part is the MANIFEST file. At the
comment, Raft storage doesn't issue a fsync command on the file. This
command is crucial to ensure that the content of the file is flushed to
the disk. Without this command, the file might be damaged and unreadable
later.