storage: Ensure MANIFEST file is properly flushed to disk
In certain situations, the manifest file is modified during transaction
processing. Currently, we are not flushing the file to disk before
admitted and afterward if it has been altered. This oversight could lead
to file corruption or absence from the log entry if a failure occurs
before it is written to disk.

We will introduce the necessary fsyncs for both the file and its parent
directory recursively to address this issue.