raft: Generalize snapshotter metrics
In raftmgr package, there are some prometheus metrics encapsulated in a
metrics object. Those metrics are used for snapshotter at the moment. In
the future, Raft package needs more metrics to track the activity of
Raft. Therefore, this commit generalizes the snapshotter-specific object
so that it can be used to store broader types of metrics.