raft: Simplify snapshotter tests
Raft's snapshotter is reponsible for generating Raft-friendly snapshots.
A snapshot is sent over the network when Raft detects a missing log
entry in replicas. Recently in the test suite, the snapshotter
materializes a transaction's snapshot from a real partition. In upcoming
commits, `storagemgr/partition` package needs to import `raftmgr`
package. That leads to cycle import problem.

This commit simplifies the snapshotter test suite. Apparently, the test
suite doesn't need to create a full-feature partition. It requires the
file-system and key-value DB utilities of the partition only.