offloading: Fix flaky TestSink_Upload_Timeout_Cancellation_And_Retry
This commit fixes the flakiness in the test
TestSink_Upload_Timeout_Cancellation_And_Retry. This test asserts the
ability of offloading utility dealing with cancellation and retrying.
Thus, timing is an essential part. We cannot substitute it for a more
reliable mechanism.

In some slow environments, it might take a while for a goroutine to be
scheduled. Recently, the test suite has had tiny headroom (< 10
millisecond). This commit adjusts some timeout values and simulating
constant backoff factors. This adjustment reduces the chance of failure
due to scheduling delays. This solution does not get rid of the
flakiness completely, though.