migrations: Cleanup dry-run manager
The dry-run mode for the migration manager was introduced to help test
the reftable migration in a dry-run state. The idea was to collect
metrics for the reftable migration.

However, there was a small oversight. Reftable migrations often take
more time than the RPC which triggered the migration. This is not a
problem for the actual migration which is blocking in nature. However,
in dry-run mode, the migrations are run in the background. But being
spawned from the RPC, the transaction is closed when the RPC finishes
due to the partition being closed. This triggers a context cancellation
often causing the migration to end mid way.

To avoid this, we're now introducing an RPC to help dry-run the
migration. So let's cleanup all the code added here since it will no
longer be utilized.