git: Move selection of reference backend to command factory
When a new repository is being created, it will be created using the
files reference backend by default. To use the reftable backend, the
'--ref-format' flag must be explicitly set.

While some of the commands do follow this, newer commands often miss
this flag. To ensure we don't miss this, let's centralize this to ensure
that all invocations of 'git-clone(1)' and 'git-init(1)' explicitly set
this flag.

While we're here, also ensure that the reftable backend is only enabled
when we have transactions enabled. Since without transactions,
modifications to the reftable backend can cause deadlocks.