repository: Don't check for flag in `DryRunReftableMigration`
The `DryRunReftableMigration` RPC is used to dry-run the reftable
migration. The RPC is invoked directly from within the pod. This means
there are no featureflags set (since we're not going through rails). So
let's remove the check for the flag here. This is okay since this RPC is
only invoked manually and doesn't raise any risks since it is not
integrated with rails.