migration: Make the reftable backend migration generic
The reftable backend migration code is specific to migrating from the
files backend to the reftable backend. It would be a lot more beneficial
if this migration allowed migration both ways, helping us revert as
necessary whenever the situation arises.

To do so, modify the migration to:
1. Receive the reference backend to migrate to, so a user can use this
migration to migrate to either the 'files' or the 'reftable' backend as
necessary. Modify the code within the migration to work as necessary
with these changes.
2. Receive the featureflag to disable/enable the migration. Instead of
hard-coding the featureflag used, the user can provide the necessary
featureflag.
3. For the files backend, pack the refs if the version of Git is less
than 2.48.