grpc: Remove FailOnNonTempDialError options
A subsequent commit will use `grpc.NewClient` instead of the deprecated
`grpc.DialContext` function to create a client connection.

One caveat with switching to `grpc.NewClient` is it doesn't support the
grpc.WithBlock and grpc.FailOnNonTempDialError options, as unlike
`grpc.DialContext`, it doesn't perform any I/O until the connection is
actually used.

Those options were originally added [1] to prevent `grpc.DialContext`
from blocking indefinitely when given an invalid address, and are no
longer needed because blocking will no longer occur during the
instantiation of the client connection.

Remove the options and the associated test case. The options are no-ops
when `grpc.NewClient` is used. We don't remove the function from the
public `client` package, as it may still be used by consumers of the
package; we simply return an empty slice instead.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2849