test: Bump tmpfs size
Occasionally, we observe random test failures due to "no space left on
device" error. The root cause of that failure is due to tmpfs runs out
of predefined space. This commit bumps the limit size slightly.

In theory, we can bump the limit more. Tmpfs does not occupy the
predefined size at mounting time. It lifts up the memory usage along
with real tmp directory usage. When the memory is under pressure, it
offloads a portion to swap. Thus, the limit should not affect the runner
resource usage.