test: Introduce GOTESTSUM_OPTIONS
This commit introduces a way to tweak the behavior of gotestsum. Some
pipelines, such as FIPS ones, need to be more forgiving of failures.
They run on a harsh, restrictive environments, therefore more sensitive
to flakiness. This commit also increases the maximum failure setting of
gotestsum autoretry for FIPS jobs.