raft: Fix racy committedLSN access
raftmgr.Storage maintains a committedLSN value. It is essentially a
derived value of etcd-raft's hard state. It is updated after the hard
state is persisted. In the current implementation, raftmgr.Manager
accesses committedLSN freely, causing a race with is write path, which
is protected by a mutex. This commit fixes that race by protecting the
read path with a mutex.