Bundle-URI: Fix inconsistent state key
The current strategy used to generate bundles
has an internal in-memory state that keeps
track of how many time a clone request was
received, and if a bundle is currently being
generated for this particular repo.

The current key used for that state is a path
formed of the storage path + the relative repo
path. This key pattern leads to inconsistencies
when transactions are enabled, because the storage
path includes the snapshot path.

To make matter more simple, this commit changes
the key used to only include the Gitlab repo's ID.
This path is immutable so it will be constant
thoughout the lifecycle of the process.