raft: Fix flaky LSN offset in tests
Occasionally, the test suite (with Raft enabled) fails due to mismatched
snapshot LSN. The failure happens in arbitrary tests.

```
=== NAME  TestTransactionManager/a_random_test
    testhelper_test.go:1192:
                Error Trace:    /internal/gitaly/storage/storagemgr/partition/testhelper_test.go:1192
                                /internal/gitaly/storage/storagemgr/partition/transaction_manager_test.go:390
                Error:          Not equal:
                                expected: 0x1
                                actual  : 0x2
                Test:           TestTransactionManager/a_random_test
                Messages:       mismatched ExpectedSnapshotLSN
```

When Raft is enabled, etcd/raft engine inserts internal log entries
(conf change and verify entries) alongside with application-issued
entries. The insertion point of time is non-deterministic. Thus, those
insertions break LSN assertion in most WAl tests.

Recently, raftmgr.Manager records all inserted entries, including both
internal ones and application-issued ones. In the test teardown phase,
the expected LSN is offset to compensate for LSN slots occupied by
internal ones.

For example, there are 3 recorded entries: R R A. `R` denotes entries
from etcd/raft, and `A` denotes the entries from the application. A test
asserts the applied LSN is 1, which is correct for a non-Raft setting.
When Raft is configured, the applied LSN must be 2 or 3
(non-deterministically). The asserted LSN 1 is shifted to 2/3
respectively.

The current implementation of the entry recorder is not correct. It
returns an incorrect result if there is no application log entry. For
example, "R R". LSN 0 is shifted to 1 instead of 2.

After fixing the offset implementation, LSN 0 is now always shifted to 2
in the "R R" case. Unfortunately, this fix exposes a bigger problem. The
snapshot creation depends on a random committed LSN. Two cases may
occur:
- R [Start transaction] R [Commit transaction] A. The correct snapshot
  LSN must be 1.
- R R [Start transaction] [Commit transaction] A. The correct snapshot
  LSN must be 2.

There's no way to differentiate two cases. The point of time when a
transaction starts is not predictable.

To overcome this problem, this commit refactors offsetting. The event
registry implements two offset-shifting methods:
- OffsetRight: the fixed version of the current Offset implementation.
  It is not used for assertions in test teardown.
- OffsetLeft: excludes internal Raft entries. Instead of shifting the
  expected LSN, tests now shift the actual LSN. This method is used in
  assertions when the test suite is running.

In the above cases, the offset snapshot LSNs are both 0, which matches
the non-Raft setting. Two offset methods cannot be merged. Some
assertions assert the additional log entries from Raft. Thus, we need
both of them.