raft: Remove `metadata` field from log entry proto
The `metadata` field was introduced in early iterations. That field was
supposed to store Raft-related metadata, especially Term. We didn't move
forward with that direction. Log entries are considered to be pure data.
They should be detached from clustering information. Therefore, the
current implementation stores Raft-related info in the key-value
database.

This commit removes that field from log entry schema. This field has
never been set anywhere in the code. As a result, it's safe to remove it
without the need for reservation.