gitaly: Allow more than one transaction interceptors
To transactionalize requests, we add a new transaction
interceptor (unary and stream). The interceptor is used to intercept the
request and add a transaction around it.

In the following commit, we want to introduce a new interceptor for
oppurtunistic migrations. This interceptor should be a layer below the
transaction interceptor. It would take transactionalized requests and
queue a migration request or a migration cancellation based on the type
of the request. To allow this, let's make `server.TransactionMiddleware`
take an array of unary and stream interceptors. This will allow us to
plug in our interceptor in the following commit.