reftable: Add a new opportunistic migrator
To perform reftable migrations, initially we had envisioned that we
could use the migration framework built over the transaction manager.
However since the migration framework is a blocking operation, this
means that for repositories with large number of references, this would
block access to the repository while the migration is being performed.

To avoid a bad customer experience, we decided to add an opportunistic
migrator. Similar to the migration framework, migrations here would be
triggered by incoming RPCs on a repository. When an incoming RPC is of
'ACCESSOR' type, we trigger a migration request. This is a non-blocking
request which tries to perform the migration in the background and
commit it.

If there is an incoming write request on the same repository, we then
will cancel the migration. This allows no degradation in the customers
experience. The migration states are then stored on memory to ensure
we can add exponential backoffs to failed migrations and also track
successful ones.

Finally the migrator only performs one migration at a time to ensure
that we don't overwhelm the server and take things slowly. Since
migration registration is non-blocking, incoming requests while the
migrator is busy will be dropped.

The whole mechanism is made possible by middlewares which are behind the
`ReftableMigration` featureflag.