prafect: Handle backup RPC routing for missing repositories
We removed the repository existence check from BackupRepository
RPC, which means backup operations do not return skipped error.
When the repo doesn't exist, BackupRepository RPC would still
execute the backup code in order to write a manifest.
For gitaly standalone it works this way, but in praefect the RPC
in the gitaly node never gets called because praefect needs the
repository to exist in order to route the RPC.
Here we are adding special handler for the BackupRepository RPC
to route them storage scope to any available healthy node if the
repository does not exists.