PreReceiveHook: Chunk error output
When the call to `manager.PreReceiveHook` returns
an error that is not of type `*exec.ExitError`, the
error output is sent to the client un-chunked. This
is causing an issue on the gitaly-hooks binary when
the output exceeds 4Mb, which is the default maximum
message size accepted by the client.

This change will chunk the output response in such
circumstances, allowing the client to process the
output.

Relevant resources:
* https://gitlab.com/gitlab-org/gitaly/-/issues/5048