Makefile: make `-Dprefix=` part of Meson build options
We pass `-Dprefix=${GIT_PREFIX}` to Meson so that it knows to install
the Git distribution into the correct prefix. When that prefix changes
though we don't know to rebuild because we don't print the variable
contents into the "git-*.version" file that is used to keep track of any
changed build options.

Move the option into `GIT_MESON_BUILD_OPTIONS`. On the one hand this
makes us detect changes to the `GIT_PREFIX` variable for free. On the
other hand this also prepares for the next commit, where we introduce
logic to build the (non-bundled) Git distribution, which would otherwise
have to duplicate the `-Dprefix=` option.