middleware: Handle repository not found in transaction
Currently transaction middleware returns error if repository does not
exists to avoid creating partition assignments for incorrect repository
accesses that do not actually point to a repository.
But BackupRepository RPC allows repository to not exists, and simply
creates empty manifest file for such cases to prevent issues during
RestoreRepository. Although the RPC works this way, when transactions
are enabled it will return NotFound before hitting the RPC itself.
So here we are introducing a solution where certain RPCs will fallback
to nonTransactionalRequest when begin transaction fails due to missing
repository.