raft: Upgrade raft package to 3.6.0
This commit upgrades the raft package to 3.6.0. The most significant
change of this release is that the `raft` package release was decoupled
from the etcd release [1]. The package location was moved from
`go.etcd.io/etcd/raft/v3/raftpb` to `go.etcd.io/raft/v3`.

This migration does not affect any functionality of the existing
package. Instead, it is a plain renaming upgrade. The package version
continues from the original location.

Apart from that, in the new 3.6.0 version, the Raft lifecycle was
modified so that it's more likely for the list of committed entries to
be fed in the next ready handling.

- Before (more likely): Persist entries -> Handle committed -> Advance
- After (more likely): Persist entries -> Advance -> Handle committed ->
  Advance

This change was not technically a semantic change. Applications were
advised not to assume specific data contained in a readiness package.
Therefore, there shouldn't be a functional change at the application
layer. However, some test setups of the raft manager were impacted. They
simulate crashing events using a hook system. This new version makes the
hooks trigger earlier than expected. This commit addresses this issue by
adding some conditions to the hook.

[1] https://github.com/etcd-io/etcd/issues/14713