raft: Add DefaultFactoryWithNode to facilitate manager registration
We need a bidirectional reference to raft-enabled storage and raft
manager. However, because we wrap the Node with Raft functionality after
it’s instantiated, we lack direct access to the raft-enabled storage
when creating the Manager. The DefaultFactoryWithNode resolves this by
registering the newly created Manager with the storage and setting the
storage reference on the Manager.
Later, the raft-enabled storage will be leveraged by the raft manager to
retrieve routing table and transport.