deps: Update go-grpc to 1.71.1
We recently [1] started instantiating gRPC client connections using the
`grpc.NewClient()` constructor rather than the deprecated
`grpc.DialContext()` function. This was primarily done to resolve an
issue where hostnames couldn't be used in conjunction with the tls://
scheme, as there was no way to specify that external DNS resolution was
required.

The move to `grpc.NewClient()` has appeared to cause problems when the
configuration includes the use of the http_proxy, https_proxy, and
no_proxy environment variables, and also with certain TLS certificates:

- https://gitlab.com/gitlab-org/gitlab/-/issues/526308
- https://gitlab.com/gitlab-com/request-for-help/-/issues/2604
- https://gitlab.com/gitlab-com/request-for-help/-/issues/2599

This upstream issue [2] describes the problem with the use of no_proxy;
`grpc.NewClient()` appears to be prematurely resolving provided
hostnames into IPs, causing the resulting address to not match with
the hostname defined in no_proxy.

The TLS certificate issue could also be related, but there are no
upstream issues filed specifically for that one.

Update the go-grpc dependency to the latest version, which contains the
fix [3] for the no_proxy issue.

The opentelemetry package appears to be contained within grpc-go v1.71.1
and thus causes an ambiguous import error:

> go mod tidy
go: gitlab.com/gitlab-org/gitaly/v16/internal/backup imports
	gocloud.dev/blob/gcsblob imports
	cloud.google.com/go/storage imports
	google.golang.org/grpc/stats/opentelemetry: ambiguous import: found package google.golang.org/grpc/stats/opentelemetry in multiple modules:
	google.golang.org/grpc v1.71.0 (/Users/james/go/pkg/mod/google.golang.org/grpc@v1.71.0/stats/opentelemetry)
	google.golang.org/grpc/stats/opentelemetry v0.0.0-20240907200651-3ffb98b2c93a (/Users/james/go/pkg/mod/google.golang.org/grpc/stats/opentelemetry@v0.0.0-20240907200651-3ffb98b2c93a)

We can remove the indirect dependency from the go.mod.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7673
[2] https://github.com/grpc/grpc-go/issues/7556
[3] https://github.com/grpc/grpc-go/pull/7881

# Conflicts:
#	go.mod
#	go.sum