grpc: Disable ALPN when TLS is used
Newer versions of grpc-go attempt to use ALPN by default when TLS
addresses are used. This causes problems with some customer
configurations of Gitaly, and thus is a breaking change in our context.

An experimental constructor [1] was introduced for creating TLS
credentials that don't use ALPN. Use this by default to retain the
previous behaviour of grpc-go.

[1] https://github.com/grpc/grpc-go/pull/7980